fix: second-round review feedback
- Propagate audience sync failures from the preferences API instead of
  reporting success when the Resend unsubscribe/resubscribe fails
- Drip emails link per-category unsubscribe (cron + immediate send path)
- Extract SSRF guards into lib/utils/safe-fetch and apply them to the
  AI Coach fetch-job-description route (blocked hosts, DNS rebinding
  check, 10s timeout, 2MB response cap)
- Enforce the size cap on readBodyCapped's no-stream fallback
- Escape ctaButton text/url and the wrapEmail unsubscribe link in shared
  email templates, mirroring the transactional templates
